requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncstdlib==3.10.9
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
curl-cffi==0.5.10
docx==0.2.4
duckduckgo-search==3.9.9
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
ezodf==0.3.2
fastapi==0.104.1
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.0
frontend==0.0.3
frozenlist==1.4.0
future==0.18.3
g4f==0.1.9.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
mmh3==4.0.1
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
Pillow==10.1.0
platformdirs==4.0.0
prov==2.0.0
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
PyExecJS==1.5.1
pyjsparser==2.7.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
pyth==0.6.0
python-dateutil==2.8.2
python-docx==1.1.0
python-pptx==0.6.23
pytils==0.4.1
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
requests==2.31.0
scipy==1.11.4
selenium==4.15.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
striprtf==0.0.26
tools==0.1.9
traits==6.3.2
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.24.0.post1
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
XlsxWriter==3.1.9
yarl==1.9.3
